energy_consumption_monitor/requirements (1).txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
paho-mqtt==1.6.1
python-json-logger==2.0.7
prometheus-client==0.19.0
aiofiles==23.2.1
python-multipart==0.0.6
# ML and anomaly detection
scikit-learn==1.3.2
numpy==1.26.2
joblib==1.3.2
# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
